visning av kontakter per kategori

+++ app/src/main/java/com/example/lab8_3/ContactRepository.java
@@ -37,7 +37,7 @@ public class ContactRepository {
  //. . .
     }
 // . . .
-    public LiveData<List<Contact>> getContactsByCategory(int categoryId) {
+    public LiveData<List<Contact>> getContactsByCategory(long categoryId) {
         return contactDAO.getContactsByCategory(categoryId);
 
     }
